Fix app crash
