+++ client/.eslintrc.json
@@ -1,27 +1,59 @@
 {
-  "extends": ["airbnb", "prettier/react", "plugin:react/recommended"],
-  "plugins": ["react", "prettier"],
-  "rules": {
-    "prettier/prettier": ["error"],
-
-    "consistent-return": "off",
-    "import/no-unresolved": "off",
-    "import/prefer-default-export": "off",
-    "no-param-reassign": "off",
-    "operator-linebreak": "off",
-    "object-curly-newline": "off",
-
-    "react/jsx-props-no-spreading": "off",
-    "react/require-default-props": "off"
-  },
-  "parser": "babel-eslint",
+  "root": true,
   "env": {
     "browser": true,
-    "jest": true
+    "jest/globals": true
+  },
+  "extends": [
+    "airbnb",
+    "plugin:@typescript-eslint/recommended",
+    "plugin:prettier/recommended",
+    "prettier/@typescript-eslint"
+  ],
+  "parser": "@typescript-eslint/parser",
+  "parserOptions": {
+    "project": "./tsconfig.json",
+    "sourceType": "module"
+  },
+  "plugins": ["@typescript-eslint", "jest"],
+  "rules": {
+    "import/extensions": [
+      "error",
+      "ignorePackages",
+      {
+        "js": "never",
+        "jsx": "never",
+        "ts": "never",
+        "tsx": "never"
+      }
+    ],
+    "import/prefer-default-export": "off",
+    "import/no-extraneous-dependencies": "off",
+    "global-require": "off",
+
+    "no-param-reassign": [2, { "props": false }],
+    "spaced-comment": ["error", "always", { "markers": ["/"] }],
+    "consistent-return": "off",
+
+    "react/prop-types": "off",
+    "react/jsx-filename-extension": [1, { "extensions": [".jsx", ".tsx"] }],
+    "react/jsx-props-no-spreading": "off",
+    "react/button-has-type": "off",
+
+    "@typescript-eslint/no-var-requires": "off",
+    "@typescript-eslint/explicit-module-boundary-types": "off"
   },
   "settings": {
+    "import/extensions": [".js", ".jsx", ".ts", ".tsx"],
+    "import/parsers": {
+      "@typescript-eslint/parser": [".ts", ".tsx"]
+    },
+    "import/resolver": {
+      "node": {
+        "extensions": [".js", ".jsx", ".ts", ".tsx"]
+      }
+    },
     "react": {
-      "pragma": "React",
       "version": "detect"
     }
   }
