+++ client/package.json
@@ -10,6 +10,7 @@
     "@types/react": "^16.9.35",
     "@types/react-dom": "^16.9.8",
     "@types/react-redux": "^7.1.9",
+    "@types/react-router-dom": "^5.1.5",
     "@types/webpack-env": "^1.15.2",
     "antd": "^4.2.2",
     "axios": "^0.19.2",
@@ -37,7 +38,7 @@
     "test": "react-app-rewired test --watchAll=false --coverage --detectOpenHandles --forceExit",
     "styleguide": "styleguidist server --config styleguide.config.js",
     "styleguide:build": "styleguidist build --config styleguide.config.js",
-    "lint:fix": "eslint './src/**/*.{js, jsx}'"
+    "lint:fix": "eslint \"src/**/*.ts*\" --fix"
   },
   "engines": {
     "node": ">= 12.0.0",
@@ -60,6 +61,8 @@
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.5.0",
     "@testing-library/user-event": "^7.2.1",
+    "@typescript-eslint/eslint-plugin": "^3.3.0",
+    "@typescript-eslint/parser": "^3.3.0",
     "babel-eslint": "^10.1.0",
     "babel-plugin-import": "^1.13.0",
     "eslint": "^6.8.0",
@@ -67,9 +70,9 @@
     "eslint-config-babel": "^9.0.0",
     "eslint-config-prettier": "^6.11.0",
     "eslint-plugin-import": "^2.20.2",
+    "eslint-plugin-jest": "^23.16.0",
     "eslint-plugin-prettier": "^3.1.3",
     "eslint-plugin-react": "^7.19.0",
-    "eslint-plugin-standard": "^4.0.1",
     "husky": "^4.2.5",
     "prettier": "^2.0.5",
     "react-styleguidist": "^11.0.5",
