+++ client/src/components/Button/Button.types.ts
@@ -1,3 +1,5 @@
+import React from 'react';
+
 type AttributeProps = React.ButtonHTMLAttributes<HTMLButtonElement>;
 
 type CustomProps = {
