+++ client/src/components/Sample/ReduxSample.types.ts
@@ -1,7 +1,7 @@
 export interface UserSampleState {
   id: number;
   email: string;
-  first_name: string;
-  last_name: string;
+  firstName: string;
+  lastName: string;
   avatar: string;
 }
