+++ client/src/components/index.ts
@@ -1,1 +1,2 @@
+/* eslint-disable import/extensions */
 export { default as Button } from './Button/Button';
