+++ client/src/store/ReduxSample/ReduxSample.types.ts
@@ -1,8 +1,8 @@
 export interface UserSampleState {
   id: number;
   email: string;
-  first_name: string;
-  last_name: string;
+  firstName: string;
+  lastName: string;
   avatar: string;
 }
