+++ client/src/store/reducers.ts
@@ -1,0 +1,8 @@
+import { combineReducers, Reducer } from '@reduxjs/toolkit';
+import reduxSampleReducer from './ReduxSample/ReduxSample.reducers';
+
+const rootReducer: Reducer = combineReducers({
+  sample: reduxSampleReducer,
+});
+
+export default rootReducer;
